Changed exp given to new MoveTowards script
